Reduce memory usage of ExportToFile
Using `File.WriteAllText` and `Export` builds the entire CSV string and
then writes it to disk. This essentially doubles the peak memory usage
of the class and is impractical for larger CSV documents.

Using `File.WriteAllLines` and `ExportToLines` on the other hand only
keeps a single row in memory at a time.